arruma nos arquivos
