+++ poke/src/Pages/DetailsPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useHistory } from "react-router-dom";
-export const DetailsPage = () =>{
 
+export const DetailsPage = () =>{
     const history =useHistory()
     const voltar=()=>{
         history.goBack('')
